requirement.txt
pillow==10.4.0
PyAutoGUI==0.9.54